Implement basic input.
The player can now move around with HJKL.  The game advances one frame
whenever there is a keyboard input.  If there is no input, the player's
direction will not change and they will continue in the direction they
were already moving.

Moving over the apple will cause the player to grow and the apple to
reset.  A debug message displaying the player's length is output to aid
in being able to see how long the player actually is.

The game can be exited by pressing 'q' or paused by pressing 'p'.