Use more efficient drawing code.
The `draw_snake` function is used for redrawing the whole snake after a
screen wipe and becomes less efficient as the snake gets bigger.

This commit uses a more efficient method of drawing the snake by only
drawing the head each frame, and erasing the tail if and only if the
player has not eaten an apple.