Fix terminal not echoing input after quitting.
This commit fixes an error that causes the user's terminal to not echo
user input after quitting the game.  This error did not present itself
if the player died rather than quitting for reasons which will keep me
up at night, but this commit fixes the occasions where it did.