Rework player name input.
The default "anon" name now only appears if the player enters a blank
name.  Otherwise, the "name" array is filled by the player's input.
This was done to avoid an exploit whereby the player can enter a blank
name, causing the game to take in garbage input the next time it
attempts to read the hiscore file.